refactor: relocate common utilities to internal package
