--- x/dwn/types/attns.go
@@ -1,1 +1,0 @@
-package types
