--- x/dwn/types/ipfs.go
@@ -1,7 +1,0 @@
-package types
-
-import "github.com/sonr-io/core/common"
-
-func NewIPFSClient() (common.IPFS, error) {
-	return common.NewIPFS()
-}
